Extract errors & validation functions to separate files #758
