+++ src/app/import/utils/emails/error.ts
@@ -1,0 +1,4 @@
+// TODO:
+// InvalidPersonId
+// InvalidEmail
+// PersonNotFoundError
